# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-bill API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/engagements": {
      "post": {
        "operationId": "secu_engagement_create",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Create engagement (optional convenience: include primaryDomain to bootstrap entity + auth)",
        "responses": {
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "solo_lab",
                      "ctf",
                      "bug_bounty",
                      "customer_pentest",
                      "internal"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "planning",
                      "active",
                      "paused",
                      "completed",
                      "archived"
                    ]
                  },
                  "scopeSummary": {
                    "type": "string",
                    "maxLength": 8192
                  },
                  "primaryDomain": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256,
                    "pattern": "^[a-zA-Z0-9.-]+$"
                  }
                },
                "required": [
                  "name",
                  "kind"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "get": {
        "operationId": "secu_engagement_list",
        "tags": [
          "secu-engagements"
        ],
        "summary": "List engagements",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "solo_lab",
                "ctf",
                "bug_bounty",
                "customer_pentest",
                "internal"
              ]
            }
          },
          {
            "name": "ownerUserId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/engagements/:id": {
      "get": {
        "operationId": "secu_engagement_get",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Get engagement (with embedded graph snapshot + counts)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "patch": {
        "operationId": "secu_engagement_update",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Update engagement",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "planning",
                      "active",
                      "paused",
                      "completed",
                      "archived"
                    ]
                  },
                  "scopeSummary": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 8192
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "delete": {
        "operationId": "secu_engagement_archive",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Archive engagement (soft delete)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/engagements/:id/graph": {
      "get": {
        "operationId": "secu_engagement_graph",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Cytoscape-compatible graph snapshot for an engagement",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/engagements/:id/entities": {
      "get": {
        "operationId": "secu_engagement_entities_list",
        "tags": [
          "secu-engagements"
        ],
        "summary": "List entities linked to an engagement",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asset_domain",
                "asset_subdomain",
                "asset_ip",
                "asset_host",
                "asset_url",
                "person",
                "organization",
                "location",
                "credential_ref",
                "document"
              ]
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      },
      "post": {
        "operationId": "secu_engagement_entity_link",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Link an entity to an engagement (or upsert+link in one call)",
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "upsert": {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "asset_domain",
                          "asset_subdomain",
                          "asset_ip",
                          "asset_host",
                          "asset_url",
                          "person",
                          "organization",
                          "location",
                          "credential_ref",
                          "document"
                        ]
                      },
                      "primaryValue": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 512
                      },
                      "displayName": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "discriminator": {
                        "anyOf": [
                          {
                            "type": "string",
                            "maxLength": 256
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "data": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "kind",
                      "primaryValue"
                    ],
                    "additionalProperties": false
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "primary_target",
                      "in_scope",
                      "out_of_scope",
                      "pivot",
                      "context"
                    ]
                  },
                  "notes": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 4096
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
    "/engagements/:id/entities/:entityId": {
      "delete": {
        "operationId": "secu_engagement_entity_unlink",
        "tags": [
          "secu-engagements"
        ],
        "summary": "Unlink an entity from an engagement",
        "responses": {
          "204": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "data": {
                      "anyOf": [
                        {},
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0
            }
          }
        ],
        "security": [
          {
            "frontendBearer": []
          }
        ],
        "x-transport": "http"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "xApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "cronJobBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "CRON_SECRET"
      },
      "frontendBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "companyApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  }
}